mem2reg pass works properly and llc can generate asm from llvm ir properly

+++ java/utils/opt/Optimizer.java
@@ -247,7 +247,7 @@ public final class Optimizer
 
         // TODO InlinePass
         createStandardModulePasses(pm,
-                (byte)1/*optimization level*/,
+                1/*optimization level*/,
                 false/*optimize size*/,
                 false/*unroll loop*/,
                 null/*inline pass*/);
